+++ src/BracketStyle.ts
@@ -1,1 +1,0 @@
-export type BracketStyle = "allman" | "one-true-brace";