+++ src/extension.ts
@@ -1,6 +1,5 @@
 import * as vscode from "vscode";
 import { splitJoin } from "./splitJoin";
-import { BracketStyle } from "./BracketStyle";
 
 export function activate(context: vscode.ExtensionContext) {
   const disposable = vscode.commands.registerCommand("editor.splitJoin", async () => {
@@ -10,16 +9,13 @@ export function activate(context: vscode.ExtensionContext) {
       return;
     }
 
-    const bracketPairs: string[] = vscode.workspace
-      .getConfiguration()
-      .get("splitJoin.bracketPairs", ["()", "[]", "{}"]);
-    const maxScanLength: number = vscode.workspace.getConfiguration().get("splitJoin.maxScanLength", 0);
-    const delimiters: string[] = vscode.workspace.getConfiguration().get("splitJoin.delimiters", [";", ","]);
-    const bracketStyleSetting = vscode.workspace
-      .getConfiguration()
-      .get<Record<string, BracketStyle>>("splitJoin.bracketStyle", { "*": "one-true-brace", csharp: "allman" });
+    const config = vscode.workspace.getConfiguration();
 
-    splitJoin(editor, maxScanLength, bracketPairs, delimiters, bracketStyleSetting);
+    const brackets: string[] = config.get("splitJoin.detect.brackets", ["()", "[]", "{}"]);
+    const separator: string = config.get("splitJoin.detect.separator", ",");
+    const excludeRegions: string[] = config.get("splitJoin.detect.excludeRegions", ["()", "[]", "{}", '""', "''"]);
+
+    splitJoin(editor, { brackets, separator, excludeRegions });
   });
 
   context.subscriptions.push(disposable);